requirements.txt
mediapipe==0.10.0
pyserial==3.5